Endrer ISøker interface til Søker. Endrer OmDegProps til Props. Flytter formatertSivilstand til sivilstand.ts
