+++ src/test/utils.ts
@@ -2,7 +2,7 @@ import {
   Adresse,
   IPerson,
   IPersonDetaljer,
-  ISøker,
+  Søker,
 } from '../models/søknad/person';
 import {
   IBooleanFelt,
@@ -230,7 +230,7 @@ export const lagSivilstatus = (
 
 export const lagPerson = (
   hash?: string,
-  søker?: ISøker,
+  søker?: Søker,
   barn?: IBarn[]
 ): IPerson => {
   return { hash: hash ?? '', søker: søker ?? lagSøker(), barn: barn ?? [] };
@@ -244,7 +244,7 @@ export const lagSøker = (
   sivilstand?: string,
   statsborgerskap?: string,
   erStrengtFortrolig?: boolean
-): ISøker => {
+): Søker => {
   return {
     fnr: fnr ?? '',
     alder: alder ?? 0,
